refactor(ui): colocate prompts components into _components/
Renames the segment's local components/ folder to _components/ (private to the
route, matching Next's _ route-exclusion). The folder is imported only by its
own page.tsx (verified zero external importers across src, tests, and
e2e_tests), so this is a straight rename plus repointing that one index import;
a folder rename keeps every file at the same depth, so all internal and
relative imports are unaffected.

Grandfathered lint suppressions under the folder (16 entries) are re-keyed to
the new path with counts unchanged. No behavior change.
