+++ ui/litellm-dashboard/src/app/(dashboard)/prompts/_components/prompt_editor_view/ToolsCard.tsx
