+++ ui/litellm-dashboard/src/app/(dashboard)/prompts/_components/prompt_table.tsx
