+++ ui/litellm-dashboard/src/app/(dashboard)/prompts/_components/variable_textarea.tsx
