+++ ui/litellm-dashboard/src/app/(dashboard)/prompts/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import PromptsPanel from "./components";
+import PromptsPanel from "./_components";
 import { DeprecationBanner } from "@/components/DeprecationBanner";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
